modify flow direction in throughput-test script
This changes the script to send from the host on which it is invoked to
all the others, instead of sending from the first remote to all the
others.  This matches the default flow direction of iperf and is somehow
more intuitive.  Note that there is no difference in outcome if all
machines.

It does change the output displayed while it is running: one now sees
the publisher's output, which is a tad less interesting.  If everything
operates it should, then the interesting part really is in the log files
anyway.

Signed-off-by: Erik Boasson <eb@ilities.com>