Share receive buffer size in discovery
Especially for retransmits, it makes sense to try to avoid overshooting
the receive buffers on the remote nodes.  This requires them to be
communicated via discovery and for each writer to track what the minimum
is across all matched readers.

The parameter used to share this information is a vendor-specific one.

Signed-off-by: Erik Boasson <eb@ilities.com>