Limit burst sizes, pack fragments in DATAFRAG
This takes the discovered receive buffer sizes and computes a limit for
the size of the initial transmission of large samples and that of
retransmit bursts based on this (in a rather simple-mindedly fashion).

It also packs multiple fragments in a single DATAFRAG, reducing the
overhead (both bandwidth and processing).

The two are functionally independent, but the changes are intertwined.

Signed-off-by: Erik Boasson <eb@ilities.com>