wip: avoid addrset for reader-specific locators
The primary reason being that addrsets get exploded to multiple
interfaces for multicast; the secondary being that we really aren't
using anything beyond the plain locators for joining/leaving ASM groups
or for publishing discovery data.

This temporarily converts the lists in an addrset to avoid having to
update the publishing of SEDP messages before being able to try it out.

Signed-off-by: Erik Boasson <eb@ilities.com>