Allow a sertype to request keyhashes in messages
This commit adds a vendor-specific parameter to the discovery parameters
that allows a reader to request the writers generate a "keyhash", also
implementing the mechanics of letting a sertype determine whether this
is set, tracking the number of proxy readers requesting it in a writer,
and generating the keyhash if at least one reader requests it (or if the
old global override is set, or if DDS security requires it be
generated).

Making the keyhash (if included) available to serdata_from_ser is the
subject of another commit (arguably an independent one, as the keyhash
may be available also if not requested).

There exist *some* legitimate cases of using a keyhash and that is the
reason this is added.  Anyone contemplating using them should take some
specification mishaps into consideration:

The keyhash is a 16-byte blob, constructed either as:

* the big-endian serialization of the key fields, with padding zeroed
  out and any unneeded bytes also set to zero (the order is presumably
  that in which the fields occur in the serialization of the sample;
  and there is presumably padding within the keyhash to maintain CDR
  alignment, but the spec is not perfectly clear on this); or

* the MD5 hash of the serialized form, if the serialized form cannot be
  guaranteed to be at most 16 bytes long (i.e., a key consisting of just
  a bounded string of at most 11 characters is always at most 16 bytes;
  a key consisting of an unbounded string may be longer).

The idea is that this allows a more efficient way of identifying the
instances in the reader, and (as per the DDS Security spec, 7.1.1.4) to
allow infrastructures services to implement the DDS QoS (notably
history) without knowing the type or being able to decrypt the data.
Problems are plentiful, with some obvious ones:

* The first and obvious problem for using the keyhash to identify
  instances is that there can be hash collisions once MD5 gets used;

* The quality of the keyhash if it is not the MD5 hash of the serialized
  key is abominable as a hash, necessitating further hashing on the
  reader side and neutering the potential performance advantage;

* The need to serialize the key fields twice;

for which reasons Cyclone DDS originally tried to avoid generating or
using the keyhash.  With the addition of DDS Security and DDS XTypes,
some additional issues are introduced:

* The Security spec mandates that the keyhash it is *always* in "MD5
  form" when keys are to be protected. However, it is possible for a
  reader to match both secured and non-secured writers, and in such a
  case, for topics where the keyhash ordinarily is *not* MD5, the same
  instance will be identified by different key hashes for the two
  writers;

* XTypes' notion of assignability allows different matching writers to
  have different forms of keyhashes (plain/MD5) and to have differently
  sized keys that result in a different serialization.  E.g.:

    string<10> is-assignable-to string

  and can cause the plain/MD5 confusion, and

    short is-assignable-to int

  and so a writer for {int;int} and one for {short;short} both yield
  "plain" keyhashes that are different, yet describe the same instance,
  and if the first writes {0x0101,0} and the latter writes {0x01,0x01}
  there are different instances mapping to the same keyhash.

With XTypes, the type information is available, so these cases can be
detected.  Matching both protected and unprotected writers can also be
detected, and so in those cases the keyhash needs to be ignored.  This
is not yet implemented.

In short:

* Generating keyhashes is up to the writer, and this is a Cyclone
  specific extension.

* Keyhashes cannot safely be used as proxies for key values.  At best
  one can detect problematic cases and refuse to make the keyhash
  available even when it is present in the message.

And so this feature should not be relied on except in well-controlled
environments.

Signed-off-by: Erik Boasson <eb@ilities.com>