Allow mask = 0 for read/take status operation
The documentation comment suggests it is allowed, though it said "NULL"
and didn't say what exactly that means, but in any case 0 was a rather
useless value as it would turn the operation into a no-op.

The wisdom of returning an error if the mask contains status flags not
relevant to the particular entity type is also doubtful, but if 0 is
allowed to simply mean "everything", that is not really much of an
issue.

And with that change, dds_get_status_changes becomes an alias for
dds_read_status with mask = 0.  Deprecating should be considered, as the
"get" is rather ambiguous and can just as easily be interpreted as an
innocent "getter" or as a function that also reset the status.

Signed-off-by: Erik Boasson <eb@ilities.com>