Verify an MD5 collision only affects DDSI keyhash
The DDSI protocol defines a "keyhash" that is the computed from the
big-endian CDR of the key with any padding cleared
as follows:

* if the maximum length of that blob given the key types is at most 16
  bytes and the DDS Security settings allow a plain-text keyhash, the
  keyhash is that CDR padded out to 16 bytes with zeros;

* in all other cases, it is the MD5 of the CDR blob.

That means any (*) pair of instances where the MD5 hash of the keys
happens to be the same get the same keyhash.  This is one of the reasons
Cyclone DDS doesn't touch the keyhash if it can at all avoid it.

This commit adds a test that takes a well-known collision and verifies
that the instance handles are unique and that the key hash has the
expected value.

Signed-off-by: Erik Boasson <eb@ilities.com>