Data-representation in endpoint and topic QoS
This firstly introduces a derived sertype for readers and writers,
that uses the data representation as specified in the QoS.

Secondly, this changes ensure the correct data-representation (XCDR version)
is stored in the endpoint and topic qos. Invalid combinations of
extensibility of the (nested) types used in a topic and the data-representation
set via the qos will result in a bad_parameter error when creating a topic,
reader or writer.

If no data-representation is set in the QoS, a sensible value will be chosen
by Cyclone. This commit includes some tests for use of the data representation
QoS.

Signed-off-by: Dennis Potman <dennis.potman@adlinktech.com>