Key order in extract_key_from_data
Some refactoring of the cdrstream serializer, to use correct key
field order (by the field's member id) in extract_key_from_data.
The key descriptor is extended with an index value per key field,
which indicates the index of that key field in the CDR. This allows
to efficiently extract the key from the CDR of a sample and get
a serialized key in the correct order.

Signed-off-by: Dennis Potman <dennis.potman@adlinktech.com>