Update cdrtest for new Cyclone idl compiler
Because the idl compiler in Cyclone does not include OpenSplice XML type
descriptors (like the old java based compiler did), the cdrtest that
checks interoperability between Cyclone and OpenSplice needs to be updated.

This commit adds a cmake module for finding an OpenSplice installation (e.g.
a source build of the community edition). The idl compiler from this OpenSplice
installation is used to compile the idl generated by the perl script in this
test. The output is transformed and included in the test program, so that an
OpenSplice type can be instantiated based on the XML descriptor.

Signed-off-by: Dennis Potman <dennis.potman@adlinktech.com>