No malloc in common case extracting keys from CDR
A bit of space is needed to support ordering fields different in the
serialized key from their order in the serialized sample, but this is
really just a pair of numbers for each key field.  As the vast majority
of topics has only a few key fields, we can avoid the malloc in the
common case.

Signed-off-by: Erik Boasson <eb@ilities.com>