Check reader/writer entity id kind in submessage
For all submessages containing a writerId in the header, that id must
have a kind "writer". There are no submessages for which the "unknown
id" for the writer makes sense. Similarly, the readerId fields must be
of the "reader" kind, or, for all cases other than AKNACK and NACKFRAG,
the "unknown id".

The GUID lookup functions do check that whatever is found is actually of
the correct type, but they are not supposed to be used for input
validation and assert that the GUID passed in is of the correct
kind. Passing garbage into a release build is therefore harmless, but in
debug build can cause assertion failures.

If another vendor uses different entity kinds in these locations, those
will now lead to warnings about malformed messages.

Signed-off-by: Erik Boasson <eb@ilities.com>