Suppress uninit warning in fibheap_extract_min
GCC 12's static analyzer loses track of the lazy initialization of the
array of roots in the "extract_min" operation.  Given that this function
sits in some hot loops, the obvious alternative to lazily initializing
it is not very attractive because it is sized for the worst case, which
in reality is never going to be used.

Therefore, suppressing the warning of uninitialised use appears to be
the better option.

Signed-off-by: Erik Boasson <eb@ilities.com>