Some minor changes:
* clang warnings
* clang analyzer warnings
* some questions

Don't mind the use of the word "FIXME", it is merely a convenient way to
signal that I have a remark.

I've thought about the naming a bit, the instance, the use of
"endpoint", and I think there are enough places where the endpoints are
treated the same to make it reasonable to define the interface in terms
of an endpoint.

I do think it would be slightly more elegant to have a reader and a
writer type, but I think it wouldn't help the actual code. It'd be
different if it weren't C, but that's what we are working with.