store_loaned_sample should not add to loan manager
The reference in the loan manager that got added ultimately caused a
memory leak and it was necessary only because the RHC would do
uncondionally "move" a loan during a take/read operation using loans.

By not inserting samples loaned from a PSMX plugin into the loan manager
early on, the RHC can simply add them to the reader's loan manager on
returning them to the reader.

For good measure the "move" operation has been removed in its entirety.

Signed-off-by: Erik Boasson <eb@ilities.com>