Do not insert data with PSMX writer loans in RHC
Constructing a serdata from a loaned sample in the write path needs to
allow for creating one without doing anything for serializing, or even
determining the size of the serialized representation: otherwise we can
no longer do true zero-copy in the case there is no need for the data
outside the PSMX writer endpoint.

If data from this case ends up in a RHC, then the reader history cache
suddenly has a reference to a PSMX writer loan, and those do not remain
valid after publishing the data.

The solution chosen here is to use the "serdata type conversion" logic
to also take PSMX loans into account: if there is one, copy it over into
one without the PSMX loan and insert the copy.

Making the copy requires knowing how to copy it. This simply reuses the
existing code that performs the copy via the CDR representation (with is
the only way if they are of different sertypes).  That in turn requires
the CDR to be available, and so this also serialises the data in some
more cases than before (but exactly in the same set of cases where it
used to serialise when there was only support Iceoryx).

There is therefore some room for optimisation here.

Signed-off-by: Erik Boasson <eb@ilities.com>