Move dds_reader_store_loaned_sample to dds_loan.h
The dds_loaned_sample_t type is not something ordinary applications
should have anything to do with.

It is a pointer here, and so it can be left as an incomplete type in the
API in principle.  However, those places where it is needed will
(generally) require the definition to be in scope anyway, and so moving
it out into dds_loan.h moves it nicely out of sight of regular
application code.

Signed-off-by: Erik Boasson <eb@ilities.com>