dds_read/dds_take: always provide a copy
Pointers to shared copies of data (shared across data readers, possibly
across processes if using PSMX) should only be returned by the "with
loan" variants.

This commit takes care of that and documents the behaviour in much
greater detail, especially also with respect to the use of the pointer
array that read/take use for returning samples.

Signed-off-by: Erik Boasson <eb@ilities.com>