Test read/read_wl produce private/shared copies
Read (non-_wl) must always result in private copies (as per the
documentation), so that is easy to check.

Read-with-loan may or may not return shared copies, and which of the two
happens is highly implementation dependent, both of Cyclone's core and
of a PSMX plugin. This test checks it against current behaviour, in
particular verifying zero-copy and shared loans of self-contained data
types in combination with Iceoryx.

Signed-off-by: Erik Boasson <eb@ilities.com>