Let dds_read/dds_take do real loans
With this commit, dds_read/dds_take and all variants now return pointers
to shared copies of data if this is possible (which it not always is).
As a consequence, modifying the contents of the loans will have possibly
disastrous effects, whereas before one could away with it.

If a private copy is required, the application must now allocate the
memory itself (if there are any embedded pointers, those must be null
pointers).  One must now do:

    void *sampleptr = dds_alloc (sizeof (T));
    dds_take (reader, &sampleptr, sampleinfo, 1, 1);
    ...
    dds_sample_free (sampleptr, &T_desc, DDS_FREE_ALL);

to get a private copy in heap-allocated memory. (Obviously, other
allocation/freeing routines are ok as well.)

This is a breaking change and it has not been done lightly.  The
considerations are:

- This really only breaks applications that modify "loaned" memory,
  which has always been dodgy behaviour at best, and is assumed to be
  extremely rare in application code.

- The majority of applications will now benefit from zero-copy behaviour
  where it is supported, and without having to worry about which API
  function is the appropriate one to use.

- It reduces the API surface.

Signed-off-by: Erik Boasson <eb@ilities.com>