Treat serdata of derived type as-if expected type
The "sertype derivation" mechanism exists for dealing with different data
representations and tpye coercing settings.  The code used to be very
simple: if the type of a serdata differs from that of a reader, it
requires conversion.  We convert by constructing a new serdata of the
required type from the serialized representation of the input serdata.

That means the target deserializer is expected to be able to handle the
source serializer's output.  What that implies is that it is primarily
valuable for mixing programming languages, so that a sample in C's
representation can safely be processed by C++ and vice-versa.  (The most
likely scenario.)

Given that, it is of doubtful value to make a copy if the source sertype
was derived from the target sertype, just for the purpose of generating
the correct CDR version, it assumes both sertype implementations can
handle it and, presumably, correctly convert it to an application
sample.

The operative word here is, of course, "presumably".  For the current C,
C++ and Python implementations it holds.  With that being the case,
creating a copy has no upside and is bad for performance.  Based on that
premise, this commit relaxes the check.

Signed-off-by: Erik Boasson <eb@ilities.com>