WIP: Remove need to pad out serdata to multiple-of-4
An RTPS message is a sequence of submessages, and the submessage headers must be aligned
to a multiple of 4 bytes within the message.  The serialized application data can be of
any length, and so padding may need to be inserted between application data and the next
submessage header.

The stack used to do that by requiring that payload in a serdata was padded out to a
multiple of 4 bytes, but that is quite annoying/error-prone.  This commit changes the
behaviour by making sure 4 zero bytes are available immediately before the submessage
payload.

The downside is that it now is impossible for a possible INFO_DST message to be merged in
the iovec with the submessage that it applies to. So it needs some further tinkering.