Use C23 features in CU_ASSERT if available
Change generic CU_ASSERT to CU_ASSERT_EQ/NEQ/... and use typeof+_Generic to print the
values when the assertion fails.  This also adds a CU_ASSERT_MEMEQ to check whether to
blobs are the same. If they are not, it prints a hexdump of the contents.

Remove asserts following an equivalent CU_ASSERT_..._FATAL: they were there to help
static analyzers, but the "unreachable" annotation in CU_ASSERT_..._FATAL already takes
care of that.

Signed-off-by: Erik Boasson <eb@ilities.com>